Allow passing string directly to set_sfall_return in HOOK_DESCRIPTIONOBJ
- HookScripts: track argument and return types in addition to values, by default assume all types as ints
- Set allowNonIntReturn in hook code to allow non-int return values to be accepted
- Moved get_sfall_arg and set_sfall_return opcodes to the new table

+++ sfall/Modules/Scripting/ScriptValue.h
@@ -25,7 +25,7 @@ namespace sfall
 namespace script
 {
 
-enum class DataType : unsigned long {
+enum class DataType : unsigned short {
 	NONE  = 0,
 	INT   = 1,
 	FLOAT = 2,
